requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.41.0
peft>=0.11.0
bitsandbytes>=0.45.0
accelerate>=0.28.0
trl>=0.7.9

# API and services
fastapi>=0.110.0
uvicorn>=0.29.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
openai>=1.14.0

# Utilities and data processing
numpy>=1.24.0
pandas>=2.2.0
matplotlib>=3.7.0
backoff>=2.2.0
requests>=2.31.0
tqdm>=4.66.0
datasets>=2.14.0

# Optional debugging tools
tensorboard>=2.15.0
